amend tests
